requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
starlette>=0.27.0

# AWS
boto3>=1.28.64
botocore>=1.31.64

# LLM
litellm>=1.10.0
anthropic>=0.3.0

# Storage
redis>=5.0.1

# Frontend
streamlit>=1.28.0
altair>=4.2.0
pillow>=9.0.0

# WebSockets
websockets>=12.0
simple-websocket>=1.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
requests>=2.31.0

# Utilities
python-multipart>=0.0.6
aiohttp>=3.8.5
asyncio>=3.4.3
typing-extensions>=4.5.0
loguru>=0.6.0
tqdm>=4.64.0

# Documentation
markdown>=3.4.0
